add a schema visitor to convert array properties
The properties created by the flatten transformation with "[]" to indicate an array are converted to multiple value properties with an index value so that the property names match those in the instances.

For example, if the separator is an underscore ("_") and three properties are supported for each array property, then the property with name "foo[]" is replaced by three properties "foo_1", "foo_2" and "foo_3".