fix bbox filter
For bbox queries we are using the intersection of the bbox and the spatial extent of the data to avoid coordinate transformation errors when a bbox parameter is completely outside of the domain of a projected CRS in which the data is stored. Using the minimal bounding box can then lead to surprising results in particular with point features and queries in other CRSs where features on the boundary of the spatial extent are suddenly no longer included in the result. For the purpose of the bbox filter, we do not need the minimal bounding rectangle, but we can use a small buffer to avoid those issues.