Crud (#700)
* improve jackson subtype mapping

* rename transactional to crud

* rename transactional to crud

* xtraplatform upgrades

* improve json schema modification

* crud schema support

* add schema scope query parameter

* support optimistic locking

* formatting

* fix extent configuration and computation

* handle feature changes in CRUD

* fix API definition of DELETE

* use generic etag generation for streams

* fix bbox filter

For bbox queries we are using the intersection of the bbox and the spatial extent of the data to avoid coordinate transformation errors when a bbox parameter is completely outside of the domain of a projected CRS in which the data is stored. Using the minimal bounding box can then lead to surprising results in particular with point features and queries in other CRSs where features on the boundary of the spatial extent are suddenly no longer included in the result. For the purpose of the bbox filter, we do not need the minimal bounding rectangle, but we can use a small buffer to avoid those issues.

* improve derivation of current feature etag

* support crs on post/put

* upgrade plugins

* update delete

* add missing Content-Crs header definitions

* transform MutationResult bbox to CRS84, if necessary

* update tile cache on create/update feature

On delete the spatial extent is currently not returned, so the cache cannot be updated.

* xtraplatform upgrades

Co-authored-by: Clemens Portele <portele@interactive-instruments.de>