add support for CSV output (#769)
* add support for CSV output

* align with #754 (changes to the transformation context)

* Consolidate feature encoders with flat attribute schemas

In Mapbox Vector Tiles (MVT), FlatGeobuf and CSV all attributes have simple values (string, number, boolean). FlatGeobuf and CSV also must have a fixed schema for all features of a feature type. MVT and FlatGeobuf support only a single geometry property. All three formats should include an implicit flatten transformation if none is provided in the configuration and derive a flattened schema in a consistent way. This PR adds code and new classes to build upon in these and similar feature formats.

* remove unnecessary code

The outputStream in the transformation context is not used.

Co-authored-by: Andreas Zahnen <zahnen@interactive-instruments.de>