fix synchronization conflict
The time between checking for a MBTiles cache file and its creation is short, but with multiple seeding threads two or more threads may attempt to create the same file in parallel. This led to different MbtilesTileset objects for the same file, which in turn led to exceptions due to database locks.

Also adds logging in case of problems.