Features HTML: improve hideMap behavior
Previously transformations on geometry properties had no effect. This was fixed in #719. In the Features HTML representation, no map is shown when no geometry is present in the data. Since the bugfix, no map is now shown, if the Features HTML configuration included a transformation that removes the geometry property from the response. Such a transformation does not really make sense, but is present in some existing API configurations. To support those configurations, the behavior is changed to hide the map, if the feature schema (without transformations) includes a primary geometry property.