Reduce limit value for route queries (#836)
For a routing query, the limit / page size is currently 2147483647 features/segments. Due to the optimization to fetch the query response in chunks, this results in a potentially large number of chunks. Since the route is currently constructed in memory, there are practical limits to the size of a route. The limit has been reduced to 250,000 segments for now, but a TODO has been added to review the limit.