Add option to suppress empty collections in the HTML page for the Feature Collections (#946)
For cases where the feature provider contains a larger number of types and only a subset of them have items in a dataset, it can be helpful to users browsing the dataset to only show the collections with data. This adds an option to suppress collections in the HTML page of the Feature Collections resource where the known item count is 0. If the item count is unknown the collection will be included.

Note that this only affects the HTML representation of the Feature Collections resource. The collection is still a resource in the API.