Search: Remove links from responses to ad-hoc queries (#958)
This is a first step to address #906. The self, alternate and paging links are removed for ad-hoc queries.

The limitation that ad-hoc queries do not support paging has been documented as a limitation in the building block for now until this issue has been discussed in OGC. For now, use sufficiently large values for the limit parameter to avoid the need for paging.