fix regression: IS NULL not supported by WFS provider
IS NULL is required to properly implement the `bbox` and `datetime` query parameters, which match all features where the primary geometry or instant is `null`. This has been corrected as part of the recent refactoring of the FeaturesQueryImpl class.

For WFS we simply continue to ignore the condition, i.e. drop the IS NULL predicate.