Support default values for typed query parameters
Query parameters can have default values. Some `TypedQueryParameter` classes implemented the `parse()` method already to return the default value, if no value was provided with the request. However, most classes did not handle a `null` value and `parse()` was not executed for query parameters without a value. The result was that default values had to be taken care of separately.

This has been addressed:

- `QueryParameterSet.evaluate()` now also calls `parse()` for query parameters without a value.
- `parse()` now handles `null` values for all `TypedQueryParameter` classes.