JSON-FG: Update describedby links
With PR #975 the approach to providing feature schemas has changed and ldproxy will no longer provide access to JSON Schemas for JSON-FG or GeoJSON for validation. As a consequence the JSON-FG describedby configuration option needs to be updated to remove the old links (that now lead to a 404). New configuration options have been added to configure existing schemas that can be referenced.