fix issues in the derived style for feature(s)
Hillshade and background layers were noch properly handled when a style for the GeoJSON features were derived for displaying feature(s) in the HTML representation.

In addition, if terrain was included in a default style, displaying features failed (not sure, why). Terrain is therefore removed when the GeoJSON feature style is derived.